Remove scripts/address_tools.py - For some reason, this utility file containing 2 functions was placed   into scripts/ directory, which doesn't really make any sense, and the   functions in it were each only used in a single file respectively.   This moves these functions to the files they're needed in amd removes   the address_tools.py file completely.
